Enhance MLpot context management and energy handling
- Introduced new functions for better integration of Python and Fortran MLpot registration, including `_fortran_mlpot_callback_active`, `_probe_mlpot_hybrid_energy_kcal`, and `sync_mlpot_fortran_registration`.
- Updated `assert_mlpot_user_active` and `_read_mlpot_user_energy_kcal` to ensure consistent energy readings and registration diagnostics.
- Refactored `reattach_mlpot` to handle forced reattachment scenarios, improving robustness against potential segfaults during long MD simulations.
- Enhanced logging and user feedback for energy handling processes, ensuring clarity during hybrid energy calculations.

+++ mmml/interfaces/pycharmmInterface/jax_pme_cross_monomer.py
@@ -473,9 +473,9 @@ class _CrossMonomerHostEvaluator:
         t0 = time.perf_counter() if _profile_enabled() else None
         with host_ctx():
             energy, forces = energy_forces(
-                jnp.asarray(positions_j, dtype=jnp.float64),
-                jnp.asarray(charges, dtype=jnp.float64),
-                jnp.asarray(cell_j, dtype=jnp.float64),
+                jnp.asarray(positions_j, dtype=jnp.float32),
+                jnp.asarray(charges, dtype=jnp.float32),
+                jnp.asarray(cell_j, dtype=jnp.float32),
                 jnp.asarray(i, dtype=jnp.int32),
                 jnp.asarray(j, dtype=jnp.int32),
                 jnp.asarray(cell_shifts, dtype=jnp.int32),
